refactor QuillType
